move to top level nova_testing package
