--- src/other/mock/default.nix
@@ -1,6 +1,0 @@
-{
-  pythonPackages = pkgs: with pkgs; {
-    mock-jcan = callPackage ./JCAN { };
-  };
-}
-
