+++ src/other/nova_testing/default.nix
@@ -1,0 +1,7 @@
+{
+  pythonPackages = pkgs: with pkgs; {
+    mock-jcan = callPackage ./mock_jcan { };
+    nova-pytest-framework = callPackage ./nova_pytest_framework { };
+  };
+}
+
