+++ src/ros/rover/science/science/default.nix
@@ -46,6 +46,7 @@ buildRosPackage {
     nova-camera-msgs
     teleop-modular-python-utils
     nova-science-interfaces
+    nova-pytest-framework
   ];
 
   doCheck = true;
